Added testing groups
Added the jest-runner-groups library which lets use specify, and group test cases.
/**
 * Tests the OpenAI API.
 * @group local
 */
describe('OpenAIService', () => {
For example, with this, I could run yarn test --group=local and it would run the local grouped tests. You can add multiple groups to the test cases.
In addition, using this, we can add a parameter for the GitHub Actions which is yarn test --group=-local which will then exclude the local group.
You can read up on it here.
https://www.npmjs.com/package/jest-runner-groups