graphql, and passport working together
Currently having issues with session persistence.

Sorry for the huge commit. Wanted to have it working together before I uploaded something.

Some changes need to be made, for example, when we specify typeDefs for GQL, I'm only doing it for the User Management. Will have to generalize it more.

Will figure out this auth issue.